+++ docs-src/Change-log.md
@@ -3,6 +3,7 @@
 ## 5.0.0-beta.3 ()
 
 * Allowed STOMP versions is now configurable
+* renamed `version` --> `connectedVersion`
 
 ## 5.0.0-beta.2 (2018/09/18)